Split build variant computation convention
